# fleet_management_api/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: fleet@bringauto.com
    name: BringAuto s.r.o
    url: https://bringauto.com
  description: Specification for BringAuto fleet backend HTTP API
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: BringAuto Fleet Management v2 API
  version: 3.0.0
servers:
- url: /v2/management
security:
- APIKeyAuth: []
- oAuth2AuthCode: []
...
paths:
...
  /stop:
    get:
      operationId: get_stops
      responses:
        "200":
          content:
            application/json:
              example:
              - id: 1
                name: Lidická
                position:
                  latitude: 49.204117
                  longitude: 16.606525
                  altitude: 400.25
                notificationPhone:
                  phone: "+420123456789"
              schema:
                items:
                  $ref: '#/components/schemas/Stop'
                type: array
          description: All the currently existing Stops have been returned.
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find and return all existing Stops.
      tags:
      - stop
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.stop
    post:
      operationId: create_stops
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Stop'
              type: array
        description: Stop model in JSON format.
        required: true
        x-body-name: stop
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Stop'
                type: array
          description: The Stop has been successfully created.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a new Stop.
      tags:
      - stop
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.stop
    put:
      operationId: update_stops
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Stop'
              type: array
        description: JSON representation of the updated Stop.
        required: true
        x-body-name: stop
      responses:
        "200":
          content:
            text/plain: {}
          description: The Stop has been successfully updated.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Update already existing Stop.
      tags:
      - stop
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.stop
  /stop/{stopId}:
    delete:
      operationId: delete_stop
      parameters:
      - description: ID of the Stop to be deleted.
        example: 1
        in: path
        name: stopId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            text/plain: {}
          description: The Stop with the specified ID has been deleted.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        "405":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "405"
                    message: Method not allowed
              schema:
                $ref: '#/components/schemas/Error'
          description: Method not allowed
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete a Stop with the specified ID.
      tags:
      - stop
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.stop
    get:
      operationId: get_stop
      parameters:
      - description: ID of Stop to be returned.
        example: 1
        in: path
        name: stopId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stop'
          description: The Stop with the specified ID has been found and returned.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find and return a single Stop by its ID.
      tags:
      - stop
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.stop
...
components:
...
  schemas:
    Error:
      description: Error object structure.
      properties:
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Not found
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
...
    Stop:
      description: Stop object structure.
      example:
        isAutoStop: false
        notificationPhone:
          phone: "+420123456789"
        name: Lidická
        id: 1
        position:
          altitude: 400.25
          latitude: 49.204117
          longitude: 16.606525
      properties:
        id:
          example: 1
          format: int32
          title: Id
          type: integer
        name:
          example: Lidická
          title: name
          type: string
        position:
          $ref: '#/components/schemas/GNSSPosition'
        notificationPhone:
          $ref: '#/components/schemas/MobilePhone'
        isAutoStop:
          default: false
          description: "If set to true, an Order to this Stop is always automatically\
            \ created when creating Orders for the Route containing this Stop."
          example: false
          title: isAutoStop
          type: boolean
      required:
      - name
      - position
      title: Stop
      type: object
    MobilePhone:
      description: MobilePhone Primitive structure.
      example:
        phone: "+420123456789"
      properties:
        phone:
          example: "+420123456789"
          title: phone
          type: string
      title: MobilePhone
      type: object
...
    GNSSPosition:
      description: GNSSPosition primitive structure.
      example:
        altitude: 400.25
        latitude: 49.204117
        longitude: 16.606525
      properties:
        latitude:
          default: 0.0
          example: 49.204117
          format: float
          title: latitude
          type: number
        longitude:
          default: 0.0
          example: 16.606525
          format: float
          title: longitude
          type: number
        altitude:
          default: 0.0
          example: 400.25
          format: float
          title: altitude
          type: number
      title: GNSSPosition
      type: object
  securitySchemes:
    APIKeyAuth:
      description: Authentication using an API key.
      in: query
      name: api_key
      type: apiKey
      x-apikeyInfoFunc: fleet_management_api.controllers.security_controller.info_from_APIKeyAuth
    oAuth2AuthCode:
      description: oAuth2 authorization
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/auth
          refreshUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
          scopes: {}
          tokenUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
      type: oauth2
      x-tokenInfoFunc: fleet_management_api.controllers.security_controller.info_from_oAuth2AuthCode
      x-scopeValidateFunc: fleet_management_api.controllers.security_controller.validate_scope_oAuth2AuthCode